+++ pyproximal/projection/GenericIntersection.py
