+++ pyproximal/proximal/GenericIntersection.py
